Add save_ij_tiff function and update open_ij_tiff
save_ij_tiff function is now available.

An optional parameter 'fetch_extra_metadata' is available for open_ij_tiff, which will lead to the extraction of additional useful information such as bit size and display metadata and store this information in a dictionary. This 'extra_metadata' can then be passed to the save_ij_tiff function to, for example, update / maintain the default display properties in the tiff file.