factor out downloading function
Note: This changes the behavior a bit. Previous version was opening the
tiff file and immediately unlinking the file, such that it was deleted
once tifffile closed the handle. However, this was not supported on
windows, where a blank try except was employed to hide it and also
obscure what was really happening. Now the behavior on all osses is the
same and users can employ the os-specific way to clean up their temp
folder.